requirements.txt
pyahocorasick==2.1.0
watchdog==5.0.1
pywin32; platform_system == "Windows"